Changed permissions check
Old functionality: Invoking client was able to connect any invite to any role that was below the bot's highest role with the manage guild permission, which could allow a malicious actor to generate an invite which would elevate their own permissions.  Also, the permission check was wrong; if a member has the permission to manage roles, then it has the permission to manage roles for all roles below its top role.

New functionality: Removed incorrect permission check and replaced with Permissions operators, and added check to ensure that the role being assigned by the invite is assignable by the invoking member themself.

NOTE: The invrole list command does not filter invite-role connections that assign elevated roles based on the invoking member, considering the safest way to allow the list command to be used without exposing sensitive information as such